Simplified tagging.
Before we showed a column of tags in the UI between the amount and the
description for each transaction. Now we have a single column which simple
shows whether or not the current transaction is tagged. If it is then we
highlight in green both the transaction's tag indicator and the indicator next
to the pertinent tag in the summary table. The user can use the space bar to
toggle the tag for the focused transaction.