Do draw a connection.
I thought we'd want to have some stronger indication of the relationship
between a transaction and its tag in the summary, back when we had multiple tag
columns for each transaction. Now I think we still do want a connection, which
I've implemented with a faded background between the two green cells. I think
it's nice.