I implemented stick scrolling.
I want the cursor to remain slightly above the middle of the page when
possible. This prevents one's eye from jumping up and down to find the cursor,
while also giving enough context to make sense of the situation. I set it a 1/3
of the way down the page, because that's what my wife suggested, and I love
her.