Drop back to preferred deployment method.
I've been working on Heroku Issue #43475. From what I can tell, one of the
following parts of setup.py is necessary for the Heroku slug compiler to build
something that, when deployed, has the package defined by ./ in
requirements.txt on its Python path:

    - from ez_setup import use_setuptools; use_setuptools()
    - zip_safe = False
    - name="ihazmoney" vs. "IHazMoney"

I determined this by bringing over a kitchen-sink setup.py from a known working
application (aspen-io), and whittling it down until it broke. It never broke.
My theory is that at some point the slug compiling environment was mutated in a
way that now masks the underlying issue. If that's the case, then it is
behavior I would not expect, and should be better documented if not changed.
Our understanding of this issue could be further refined by spinning up new
test apps with minimal setup.py's.