Stabilize row/column coloring.
I experimented with highlighting the row and column up to the selected cell. My
mother asked for this so she could tell what was selected. I think it looks
like a selection itself, however, which it's not. I'm going back to how it was,
which is how Google Docs (and presumably Excel) do it as well: background on
the stub and head and a border on the current cell.