I renamed the project from ihazmoney to ihasamoney.
Dowski came up with the ihazmoney name. It was the first of a bunch that we
thought up along similar lines (ohaimoney, shekeltime, etc.). None of the rest
seemed worth displacing the flash of brilliance associated with the original.
But then pjz introduced us to the "I has a money" lolcat, and the existence of
the meme and my desire to make use of it seemed sufficient to change the name.
I couldn't find an original for the meme to add our own words to.