Rough in an authentication form.
I started rebuilding the login/register form that I used on logstown. There I
had a well-abstracted JavaScript library for generating forms. Here our needs
are much simpler and I started doing it from scratch. There's a bunch of polish
that I'm having to rediscover and reimplement, however. C'est la vie.

The Sign In or Register form is not yet fully wired up. There are JSON
endpoints behind it that kind of work, but the UI doesn't respond to these
endpoints appropriately yet. It's also the case that new users won't have any
data and currently our data.py code barfs on that case. I need a separate help
card on importing/uploading, as well as the implementation of same. This is a
chunk of work and needs to be done along with the billing code.