Good progress on layout tighten-up.
I cleaned up the markup and resize function so the basic layout is sound
(afaict in Chrome, Safari, and Firefox). The scrollbars are in the right place.
We got the Help Cat button as a bonus. The color scheme is strengthened. Next
step is to tighten up scrolling, then this can be pushed.

https://ihasamoney.zendesk.com/entries/21091008-tighten-up-layout