Clean up column highlighting.
At one point I wanted to highlight all cells in the current row and column not
just the stub and head and cell itself (this was at my mother's behest,
actually). I ended up not going that route, but I had old crufty code laying
around from that attempt. Turns out that was the root of the unoptimized arrow
mashing. Gone now.