Clean up scrolling.
The usual spreadsheet behavior is to allow the user to focus on any cell. But
so far in this application we've constrained the user to focusing on cells that
actually contain data. This makes it complicated to support mouse scrolling. If
focus follows the mouse then the UI is unusably jittery as the apps scrolls
horizontally to focus on the column for that row. I've gone ahead and nixed
mouse support for now.