Make columns minimally reachable.
Vertical scrolling was borked to the point that the app was unusable. I
implemented a stop-gap measure where we keep the current column in position
one. To be evolved.

I also changed the dead mouse behavior to display on mousewheel instead of
mousemove. Moving the mouse around is a passive action on the user's part, but
scrolling isn't. We don't want them to see the dead mouse until they actively
try to use their mouse. Simply moving the mouse out of the window into a
different one is not enough intentionality to warrant the dead mouse.