New direction. Scroll tables independently.
It was too confusing to use hl to categorize, and have the category follow the
current transaction. I simplified the UI by binding df to scroll to transaction
table, and jk to scroll the categories. You can then use shift to enter a
quasi-mode where jk recategorize the current transaction. Progress!