Rename TOTAL to BALANCE. Fix for ANON.
TOTAL makes you think the category subtotals are going to sum to that number,
but this is only true with an initial balance of zero, which is very unlikely.
I'm thinking the user should be able to set their current balance at any time.
Future transactions imported will go from there and automatically update their
balance.